requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
fastapi==0.92.0
fastapi-pagination==0.11.4
h11==0.14.0
idna==3.4
Khayyam==3.0.17
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.3.3
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
